Move unit tests from executor.rs to integration tests
The src/cfg(test) module tested private internals (spawn_ready_actors,
SpawnOutcome) directly; rewritten against the public Executor API and
relocated to tests/executor.rs to match the project's convention of
keeping all tests under tests/.
