+++ ailets-rs/ailetos/src/executor.rs
@@ -728,60 +728,3 @@ impl Executor {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::dag::NodeState;
-    use crate::storage::MemKV;
-
-    #[tokio::test]
-    async fn spawn_ready_actors_detects_failed_dep_via_alias() {
-        let kv = Arc::new(MemKV::new());
-        let env = Arc::new(crate::environment::Environment::new(kv));
-
-        let failed_producer = env.add_node("failed-producer".into(), &[], None);
-        let output_alias = env.add_alias(".output".into(), failed_producer);
-        let blocked_consumer = env.add_node("blocked-consumer".into(), &[output_alias], None);
-
-        {
-            let mut dag = env.dag.write();
-            dag.set_state(failed_producer, NodeState::Terminated);
-            dag.set_exit_code(failed_producer, crate::errno::EOWNERDEAD);
-        }
-
-        let async_runtime = tokio::runtime::Handle::current();
-        let infra = ExecutorInfra::new(async_runtime, &env, None, Arc::new(Mutex::new(Vec::new())));
-        let mut actor_tasks = JoinSet::new();
-        let pending = HashSet::from([blocked_consumer]);
-
-        let outcome = spawn_ready_actors(&pending, &env, &infra, &mut actor_tasks);
-        infra.shutdown().await;
-
-        assert!(matches!(
-            outcome,
-            SpawnOutcome::FailedNodeDependency(dep) if dep == failed_producer
-        ));
-    }
-
-    #[tokio::test]
-    async fn unregistered_actor_is_terminated_with_error() {
-        let kv = Arc::new(MemKV::new());
-        let env = Arc::new(crate::environment::Environment::new(kv));
-
-        let node = env.add_node("not-registered".into(), &[], None);
-
-        let async_runtime = tokio::runtime::Handle::current();
-        let infra = ExecutorInfra::new(async_runtime, &env, None, Arc::new(Mutex::new(Vec::new())));
-        let mut actor_tasks = JoinSet::new();
-        let pending = HashSet::from([node]);
-
-        spawn_ready_actors(&pending, &env, &infra, &mut actor_tasks);
-        infra.shutdown().await;
-
-        let dag = env.dag.read();
-        let n = dag.get_node(node).unwrap();
-        assert_eq!(n.state, NodeState::Terminated);
-        assert_ne!(n.exit_code, 0);
-    }
-}
